+++ NEWS.md
@@ -1,6 +1,8 @@
-# testthat (development version)
+# testthat 3.3.1
 
 * Fixed support for `shinytest2::AppDriver$expect_values()` screenshot snapshot failing on CI (#2293, #2288).
+
+* testthat now emits OpenTelemetry traces for tests when tracing is enabled. Requires the otel and otelsdk packages (#2282).
 
 # testthat 3.3.0